+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All Notable changes to `PHP Domain Parser` **5.x** series will be documented in 
 - `Pdp\DomainInterface` interface implemented by `Pdp\Domain` and `Pdp\PublicSuffix`
 - `Pdp\Domain::getContent` returns the Domain name value replaces `Pdp\Domain::getDomain`
 - `Pdp\Domain` implements the `Countable` interface.
+- `Pdp\Domain::withPublicSuffix` returns a new Domain object with a different Public Suffix
 
 ### Fixed